Plant Table Scaffold
Created a table to hold all plants. User will select from
this table of plants when adding them to their plots.
Modified the Serializer to only show the user v1 attributes.
Changed the controller to inherit from the OpenReadController.
Removed the new and edit routes from plants.
Will need to add current_user to the controllers.